+++ homeassistant/components/zha/core/const.py
@@ -33,6 +33,10 @@ CONF_USB_PATH = 'usb_path'
 DATA_DEVICE_CONFIG = 'zha_device_config'
 ENABLE_QUIRKS = 'enable_quirks'
 
+RADIO = 'radio'
+RADIO_DESCRIPTION = 'radio_description'
+CONTROLLER = 'controller'
+
 DEFAULT_RADIO_TYPE = 'ezsp'
 DEFAULT_BAUDRATE = 57600
 DEFAULT_DATABASE_NAME = 'zigbee.db'
@@ -114,6 +118,9 @@ DISCOVERY_KEY = 'zha_discovery_info'
 DEVICE_CLASS = {}
 SINGLE_INPUT_CLUSTER_DEVICE_CLASS = {}
 SINGLE_OUTPUT_CLUSTER_DEVICE_CLASS = {}
+SENSOR_TYPES = {}
+RADIO_TYPES = {}
+BINARY_SENSOR_TYPES = {}
 CLUSTER_REPORT_CONFIGS = {}
 CUSTOM_CLUSTER_MAPPINGS = {}
 COMPONENT_CLUSTERS = {}
